+++ app/kontres/views/bookable_item.py
